post 쓰기, 저장, 수정, 삭제 접근 제한
로그인 여부 확인 미들웨어 추가
로그인 한 유저만 새 글을 쓰고 저장할 수 있음
로그인한 유저와 글쓴이 비교하여 동일한 경우만 글 수정, 삭제할 수 있음

+++ controller/posts.js
@@ -22,13 +22,15 @@ module.exports = {
 
   postPage: async function (req, res) {
     const id = req.params.id;
-    const post = await Post.findById(id);
+    const post = await Post.findById(id).populate("author", "name");
     res.render("post", { post });
   },
 
   editPost: async function (req, res) {
-    const post = await Post.findById(req.params.id);
-    res.render("edit", { post });
+    const post = await Post.findById(req.params.id).populate("author");
+    if (req.session.user.id === post.author.id)
+      return res.render("edit", { post });
+    res.render("post", { post });
   },
 
   saveEditPost: async function (req, res) {
@@ -46,7 +48,11 @@ module.exports = {
   },
 
   deletePost: async function (req, res) {
-    await Post.findByIdAndDelete(req.params.id);
-    res.redirect("/");
+    const post = await Post.findById(req.params.id).populate("author");
+    if (req.session.user.id === post.author.id) {
+      await Post.findByIdAndDelete(req.params.id);
+      return res.redirect("/");
+    }
+    res.render("post", { post });
   },
 };
